Adapt Prioritization To Observed Impediments
With tools for modeling computation, it seems surprisingly popular to encourage users to blend the code to optimize for runtime expenses with the code for required observable behavior.
This integration increases complexity for efforts to implement changes.
To reduce effort spent on such activities, it seems a good idea to learn how to better keep things simple.