Support Hosting On Azure
Test operate PuppeteerSharp on Azure: Switch away from a version of PuppeteerSharp which was observed hanging when running in an Azure web app.
Inspection with the Visual Studio remote debugger on the dotnet process revealed that a thread was not returning from the invocation of `Puppeteer.LaunchAsync`.
The azure frontend returned the following message when navigating to the test app with a web browser:
> The specified CGI application encountered an error and the server terminated the process.

Now let us see if we get a better result with the newest version of PuppeteerSharp.