Expand the CLI to support adding a client Elm app
Add an option to use an Elm app to build a client and add that to the configuration. Support specifying the subset of URLs to serve the client app from: Add integration of map.json from a local file.
Add an example client Elm app to demonstrate the functionality in combination with navigation and URL changes.