Help to see how long restoring the process state takes
Add log messages to inform about time spent restoring the process store when starting the WebHost.
The new log messages look like in the following example:
```
Begin to build the persistent process for Elm app 1FD709F7A27BF8E0C42582B5CE2A5032E6098DBC250CB3190001D228B2E74735
Begin to restore the process state using the storeReader.
Restored the process state in 0 seconds.
Completed building the persistent process.
```

This is of particular interest at development time, when the app is often restarted for a new version of the source code: The developer makes a change to the application code and wants to test it. Because restarts happen more often in this case, the time it takes to restore the app state is more important than in production where a restart happens less than weekly.

In any case, it seems useful to be able to see how much time was spent on restoring the application state. Also, inform the developer what the host is doing at the moment: Add a message to the console when starting to restore the application state and a message when this is finished.

(related: https://github.com/Viir/Kalmit/issues/10)