Support build processes in integrating projects
Offer a way to build and save the configuration to a file. I discovered this while working on DRTS Game, implementing the build process to work the same way on Windows and Linux. The method found there uses a docker build which builds the configuration file in a separate `RUN` command.