Simplify: Remove redundant functionality
We can implement CORS by adding HTTP headers in the backend Elm app, as demonstrated with the deployment of DRTS Game (CORS is used to support hosting the frontend on itch.io). Since we can do this in Elm, we don't need this functionality in the framework.

I found the following code in DRTS to implement CORS:
```Elm
    maybeHttpRequestOrigin =
        httpRequestEvent.request.headers
            |> List.filter (\{ name } -> (name |> String.toLower) == "origin")
            |> List.concatMap .values
            |> List.head

    corsResponseHeaders =
        case maybeHttpRequestOrigin of
            Nothing ->
                []

            Just httpRequestOrigin ->
                if httpRequestOrigin |> isCorsOriginAllowed then
                    [ { name = "Access-Control-Allow-Origin", values = [ httpRequestOrigin ] }
                    , { name = "Access-Control-Allow-Credentials", values = [ "true" ] }
                    ]

                else
                    []
```