Automate generating Elm type serializer 🚀
Automatically derive the functions for serializing and deserializing from the app state Elm type. Generate the Elm code and add it to the lowered Elm app before the final Elm compilation.

The introduction of these code generators saves the effort needed before to maintain the serializer code each time the state type changes. Because of the manual maintenance, each change of the state type graph implied a risk of introducing a bug into these functions. This risk was mitigated by automated tests, which had to be maintained manually as well. By formalizing and automating the process of writing the serializer code, we also obviate these kinds of automated tests. The reduction of test results in elm-test

There was a question regarding this functionality:

> How does this work with types the original Elm app code does not export from the defining module ('opaque types')?

Supporting unexported types is not necessary for the first version. Even without supporting such scenarios, the benefit is large enough to warrant releasing it as a separate version.

## How to find the root type name?

We need to get the name of the app state type from the root module. Let's look into one of the example apps to see how this type is referenced:

```Elm
interfaceToHost_processEvent : String -> State -> ( State, String )
interfaceToHost_processEvent =
    InterfaceToHost.wrapForSerialInterface_processEvent processEvent
```

By reading this functions type annotation, we learn that the name of the type is `State`.

🌕

## Pivot Control Flow

About five hours in, I had a partial implementation using this approach: First parse the whole type tree of the state type. Then build the JSON encoding and decoding functions based on that type info. Then I had the idea to simplify the whole process by removing the eager type parsing and instead let control flow from the rendering function.