Adapt the webserver interface to support migrations
Expand the webserver configuration interface to support integration in another app. One way to reach migrations is to compose an app that instantiates and contains apps like the existing app model, which does not support migrations. Migrations become possible by using the containing app that removes the more limited app instance and starts a new one for each migration. We could contain that limited app instance in a volatile host. As long as we remember the state of the process store, we can let it restore from that state.
Add an automated test to verify the `Immediately` property as explained in the guide at https://github.com/elm-fullstack/elm-fullstack/blob/1cd3f00bdf5a05e9bda479c534b0458b2496393c/guide/persistence-in-elm-fullstack.md