Simplify the configuration of web apps
Do not require mixing the API to set the Elm app state with the public interface: Provide the APIs to get and set the Elm app state in the dedicated admin interface. Adapt the automated tests to use the new interface. This change will also lead to simpler elm-fullstack.json configuration files, as apps can remove this path from the regex expression used to decide where to serve the frontend app HTML document.
Also, fix a bug in the CLI, which caused a crash when using a relative file name for the output.