Improve tooling for development scenarios
Support this use case during development: Start a new server replicating the process from a remote host (for example, from production). Continue the `deploy` option to support the application of a deployment on top of that replica.

Also, to support testing the new functionality, add an option to configure the URLs for the public app.