Fix HTTP content headers in public app web server
And enable the test that failed because of the problems with reading the HTTP response content.

I spend four hours on this, was confused by the error messages on the client-side. After switching the test implementation to use `HttpWebRequest`, I got this error message, which helped draw attention to the cause: "System.IO.IOException: The response ended prematurely."
With HttpClient, the error messages differed depending on using async/await or not. In some cases, I got: "System.IO.IOException: Unable to read data from the transport connection: The I/O operation has been aborted because of either a thread exit or an application request.."