Guide users in case of misspelled commands or options
When we get an unrecognized command or option, search one with a similar name and show it to the user.
Here is an example from a test:

```
elm-fullstack run-server  --remove-previous-process
```

This command-line now leads to the following output:
```
Specify --help for a list of available options and commands.

Unrecognized option '--remove-previous-process'

Did you mean 'delete-previous-process'?
```