Add example app with chat room
This chat app is largely based on the implementation in DRTS game. At some point, I was wondering if I should use a minimal example instead but eventually opted for a real-world example app. For now, user IDs are automatically assigned based on cookie scopes.
Some more features to add / transfer from DRTS: Choose a user name, display users online, rate-limiting for messages, play a sound when a message is added, reduce latency (possibly with long-polling).
It would be nice also to integrate an option to enable login via user accounts. The login could also be done analogous to the login process in DRTS.