Support migration of apps with short interruptions
Expand the lowering implementation to detect cases in which the module to generate the JSON coding functions was already lowered and leave it alone in these cases. This branch enables to migrate apps by merely switching to the new docker image, without requiring a new deployment with the CLI tool.