Support app codes optimizing for runtime expenses
I saw apps spending a lot of time on encoding the `Bytes.Bytes` value to base64 when building HTTP responses. As a quick way to optimize runtimes expenses, offer a base64 string directly so that apps can avoid the roundtrip to and from `Bytes.Bytes`. This should become obsolete with a better engine running the Elm code: These values could be cached, but the current engine does not do that.