Simplify compiler interface to integrate Elm make
Remove the hardcoded special module name for the source file given to Elm make. Instead, let the app specify the source file name with the same syntax as in the source files interface. Also, introduce a syntax to model optional flags in the function name.
For now, continue to support the older interface variant, to support the smooth migration of apps.
Adapt the rich-chat-room example to demo the new interface.