Begin implementing the Elm editor 🚀
Reduce friction to get started with development: So far, the recommended approach to "Setting up the Programming Tools" is relatively expensive for new people. They are asked to install an editor, an extension for the editor (for the Elm specific functionality), the Elm executable file, elm-format, and elm-analyse. Then they need to configure these to work together. For motivation, we show the benefit of seeing the Elm diagnostics in the editor. For example, the walkthrough videos on concrete projects reiterate this point.

Many people still don't make it through that setup, so we make the setup easier. The change in this commit is the first step in implementing a more accessible programming environment. People can start editing right away in a web-browser. To support scenarios of frontend apps, they can see the compiled app besides their code. For non-frontend apps, they could get a URL that describes their app to feed into other environments to run an app.