Add syntax highlighting of Elm code
Add the first version of syntax highlighting. This version uses monarch that comes with the monaco editor anyway. It looks like monarch could be sufficient colorization at least on par with the VSCode Elm extension. The most interesting parts today were getting multiline comments and the three quotes strings to work. The documentation and playground at https://microsoft.github.io/monaco-editor/monarch.html were sufficient to learn these parts.
The monarch language definition is located in a dedicated file to make it easy to copy to and from that playground. I started from the `mylang` example in the editor and copied in bits from other examples. There is probably still a bunch of stuff to remove from the monarch.js file.