Implement the first difference-based project state model 📐🚀
Reduce the effort required to persist or share a project: Reduce the length of links by encoding file contents more efficiently: Reuse contents that we already know how to load from a git repository. Describe the blob contents as the difference to that baseline from git.

When starting the quest, I had a different idea for this first iteration of the diff-based project state model. I thought to get closer to later/future models and used a combination of `List.take`, `List.drop` operations to reuse parts of blobs from the base composition. That model allows for more efficient encoding, for example of reordering a block in a file.
The version in this commit is a more pragmatic variant, cheaper to implement, thanks to work published earlier in `jinjor/elm-diff`.