Fix bug in opening a project and improve workspace state model 🐛🧹
Fix bug seen when entering a project from a URL: Do not mix the state of the file opened in the editor with the default project.
Adapt workspace state model to prevent type-level inconsistencies: Do not offer a value for an opened file path if loading the file tree failed or is not complete yet.
Avoid redundant building of commands to forward the new text to the code editor in javascript land: Use a new root-level update function to derive the necessary commands to update the parts outside the Elm app.