Reduce runtime expenses when garbage-collecting in the process store
Reduce processing time for the `truncate-process-history` command: Specialize implementation for this entry point: Separate the loading of composition events and their dependencies from their application to restore the volatile representation. The truncate-process-history method only needs the former part. The latter is not necessary in this case.

One of the instances in production with version 2021-01-16 frequently reported more than 25 seconds duration via `getFilesForRestoreTimeSpentMilli`.