Improve the state model type to prevent inconsistent states
The previous state type allowed for inconsistent combinations, such as the classical loading project failed and a compilation result. Move the fields that should be reset when loading another project into the workspace wrapper.
Reflect this structure separating the workspace from the whole app also in the Event type.

This refactoring also demonstrated how the Maybefication of the current time leads to complications. To simplify this, use a framework that provides us with the current time already on init so that we do not have to use `Maybe` here. Since this problem already appeared in other projects earlier, place that framework in the dedicated reusable module `BrowserApplicationInitWithTime`.