Migrate compiling ElmMake module to Elm-based implementation ⚙️🚀
Continue the migration to an Elm-based implementation of the app compilation: This time for the `ElmMake` module.

Implementing the `ElmMake` compilation interface is interesting because it differs from the previously migrated parts: The `SourceFiles` and `GenerateJsonCoders` were simpler design-wise as they only need the file tree as input. But the values we (re)place into the `ElmMake` module depend on the output from the Elm compilation process. That compilation was and still is done using the `make` command in the elm executable file. But the Elm modules to compile can depend on the previous compilation stages, for example, integrating JSON encoding functions that we just generated from type annotations.
How can we depend on an external program in the middle of a purely functional compilation pipeline?
This implementation uses the return value from the compilation function to tell the integrating program what dependencies were missing at that stage. The compilation function packages this information in the `MissingDependencyError` variant of `CompilationError`. The integrating software can then perform the `elm make` and retry the compilation with a dictionary of resolved dependencies as an argument.
We can reuse this pattern for other external dependencies. An example is integrating files from URLs into git repositories, similar to the `SourceFiles` interface.