Migrate compilation for migrations to new compilation pipeline
Simplify the implementation of migrations by reusing the same compilation pipeline as the compilation interface for app code.

To implement the emitting for migrations, reuse earlier implemented `mapAppFilesAndModuleTextToSupportJsonCoding`. With this new application/integration, I noticed we can simplify the implementation in and around this function, leading to refactoring to `mapAppFilesToSupportJsonCoding`: Instead of returning a function to update the module text, we now what is necessary to build (or update) that module: The names of the modules we depend on for types and the name of the generated module.

Besides what is necessary for migrations, also simplify the compilation in some places.