Fix JSON decoding of events in demo apps
Consider that popular approaches to build the JSON strings sometimes lead to adding these fields with a null value instead of omitting the field. We can see an example in the C# here using .NET types to model the structures. In many cases, we can remove the fields with default/null values from the JSON by adding an attribute on the field/property declaration, but it is easy to overlook.
Adapt the decoders on the Elm side to avoid a misunderstanding in this case.