Reduce runtime expenses with specialized file tree representations
Reduce the processing of file trees in the frontend: Introduce a specialized state model that caches more frequently used derivations.
Split functionality to work with file trees into a module for the generic functions and another one for front-end-specific parts.
Reduce runtime expenses for transport and support larger projects: Filter the files sent to the backend for compilation.

Also, improve names of file tree representation types for persistence.