Avoid process restoration failing on partially written last composition log record
Since we cannot depend on writes to the process store happening atomic, we consider scenarios in which a starting server finds a store with the last composition partially written to the representation on the file system.

Automate testing this scenario with a new test on the webserver interface.
Expand the test setup framework to support configuring the file store.
Expand the library around file stores to support setting up test scenarios and performance profiling: Add the `RecordingFileStoreWriter`