Forward commands from the init and migrate functions
Forward the commands coming out of the init and migration functions in the applications' program code.
Consolidate the compilation lowering stage for migrations into the backend module compilation: Reuse the route for other events also for init and migrate. This way, we get the commands the same way.
Expand the generate backend root module text to include functions for initialization and migration.
Also, update some names and namespaces for readability.