Fix expectations for suggested completions for import statements 🐛🧹
Adapt the expectations in automated tests, considering that some modules import the same module using different aliases.
The previous version was closer to behavior as observed with elm-format. However, looking closer at imports, the current version of elm-format seems to be not always compatible with Elm, sometimes breaking compiling code. For discussion of the issues around import statements, see:
https://github.com/avh4/elm-format/issues/379#issuecomment-409451995
https://github.com/avh4/elm-format/issues/445
https://github.com/avh4/elm-format/issues/577