Add completion suggestions for declarations in let-blocks
Expand the language service API for these scenarios: Add information about the cursor location to determine which let-declarations are in scope.
Automate a test for completion suggestions for a declaration in a let-block: Add Support modeling the editor state using a single string with a particular character sequence to mark the cursor location. Adapt the test framework to the limitations of the parsing functions: When building a test based on an editor state, simulate likely earlier versions of the currently opened file to find a parseable variant.