Expand compilation interface for listing source directory contents
Expand the API in the `CompilationInterface.SourceFiles` to allow apps to get whole subtrees.
Expand the automated tests for scenarios with more files.
I am still looking to simplify app code by (re)removing that new mapping function that cropped up in the sample app backend. Maybe move the type to the `ElmFullstack` module and add a generalized function there?