Assume we can afford to make this lookup an atom in Pine
Create a separate path for kernel functions to avoid infinite recursion: The lookup for the non-kernel names implements an error message if the requested name is not found. To distinguish this case, Pine wraps the found value in a list. To unwrap, the Elm-to-Pine compiler (currently) uses some kernel functions that would cause recursing into the lookup functionality.