Improve readability in compilation interface with record types
Begin to use record types to select file content encodings.
Adapt the `SourceFiles` and `ElmMake` interfaces to support the new method for selecting encodings. Support smooth migration of apps in production: Continue supporting the older method via flags in the declaration name for now.
Adapt the example app Elm Editor to demonstrate the new interface.
Expand automated tests to cover the new interface for source files.
Also, fix the mapping of the Elm syntax for record access in Elm Interactive. Expand automated tests to cover this as well.