Improve readability in compilation interface `ElmMake` with record types
Use record types now also to select the options for the invocation of `elm make`.
Adapt the `ElmMake` interfaces to support the new method for selecting options. Support smooth migration of apps in production: Continue supporting the older method via flags in the declaration name for now.
Adapt the example app Elm Editor to demonstrate the new interface.