Avoid distracting with superfluous completion suggestions in comments
Do not open completion suggestions when the user is typing in a comment: Automate tests of the language services module to cover such scenarios.

Add supporting infrastructure for inspecting syntax parser: Add a feature to inspect the parsed representation of the Elm syntax in the file currently open in the editor. Add a command to the Monaco command palette to invoke this view. Expand the output pane to display the JSON representation of the `Elm.Syntax.File.File`.