Simplify and reduce dependencies on web browser context
Remove unused interfaces to web browser URL. Mid-term, we might use the URL to persist application state, for example, progress with the lessons. In that case, we might introduce an additional entry point for compilation with a `Browser.application` that updates the URL accordingly. To support testing live in environments that don't support accessing the URL, we might then continue to maintain the entry point based on `Browser.document`