Ensure the user has enough time to see the hint and polish visuals and code
Do not remove the correction based on time, but continue to show it to ensure the user has enough time to read it.
After checking the answer, let the user continue with a dedicated button and the enter key: Add a global event subscription for the enter key.
Model the visuals after the practice UI of Duolingo.
Also, refactor the program code for better readability.
Also, fix a bug that caused the progress bar animation to get stuck at 99%.