Improve closing the top menu on focus loss - at least a bit
I fiddled for almost two hours with focus-related event handlers, trying to close the menu when the user clicked outside. Closing the menu still does not work when the focus goes into an iframe (Tested with Chrome). However, at least use a dedicated event to be notified when the focus goes to Monaco. With this change, we now close the menu when the user clicks into the editor.
It still does not close the menu, for example, when the user clicks into the iframe in the output pane 🤬