Clean up PineVM kernel module in Elm Interactive
Remove obsolete name for PineKernel from Pine module: This concept is specific to the Elm Interactive core library model, as referencing Pine Kernel functions now works via the recently introduced `KernelApplicationExpression`
Clean up names of kernel functions.
Refactor kernel functions to rely less on recursion when processing lists.

Moving from recursion-based towards list-based processing followed a realization this evening: The kernel/root model so far is not close to the simplest possible (tree) anyway. Moving further, we clarify the distinction from the simple model for a broader audience.
We leave translating this Pine language to a simpler one as an exercise for later.