Reduce runtime expenses in the Elm Interactive compilation step
Reduce the size of the serialized request to send to the compiler and the processing effort required for the compilation server to parse the request: Introduce a new optimized representation: Expand the request model and JSON format to support the reuse of values with references. Adapt the client to move values appearing multiple times into a dictionary and have tree nodes reference these declarations.