Improve Elm Interactive efficiency by caching environment mappings
Improve Elm Interactive efficiency and response times: Cache the representation of the environment after translating it for transport to the compiler. The environment only changes if the submission contains a declaration, so this cache saves us time after every submission that is only an expression.

Also, rename some declarations in the program code for better consistency.