Explore reducing complexity with index-based runtime lookups
The current name-based implementation of runtime lookups means we search the environment until we find the matching entry. Let's see if we can make the lookups more efficient: If we can depend on the order of items found in the application environment being known already at compile-time, we can emit the lookup using an index.

Switching the emitted code to index-based lookups at runtime means that emitting depends on determining the list of environment elements available in that context at runtime. That is an additional dependency compared with the earlier approach using a name-based lookup, where we searched through the environment until finding the matching entry. This new dependency caused a reordering of emission functions. Since the concrete emitted code depends on a particular order of elements in the application environment, we cannot reuse it in as many contexts as earlier. The new version emits an adapted version for every environment.

This commit splits the compilation into two stages, identified with the prefixes 'compile' and 'emit' in the declarations names. The 'compile' stage lowers to an intermediate representation between Elm syntax and Pine syntax. For example, it reduces the case-of expressions from Elm syntax into a combination of conditional and let-block. The references in the intermediate code still use strings as identifiers, so the mapping to an index happens in the emit stage.

This commit also simplifies return types of compilation functions by removing the component for the set of transitive dependencies. Instead of propagating the dependencies from the leaves towards the root, we now use the new intermediate expression model to derive the set of (remaining) transitive dependencies where we need to know them.

A pleasant surprise was the intermediate representation was much more effective than my intuition predicted.
The idea to introduce an intermediate representation was already months old. In this iteration, it had been frequently on my mind already days before its implementation. Intuition said it would be more difficult to implement. However, in less than four hours beginning its implementation, I reached a state where at least all elm-test-rs tests passed. Overall, it has greatly reduced the mental load when working on the compiler.