Support building static websites with `blobMain` compilation root
Add a new kind of compilation root to use with the `make` command: `blobMain` is of type `Bytes.Bytes`, and the CLI writes these to the given output file name.
Using a library like agu-z/elm-zip, we can also use the blob entry point to model a directory structure with multiple files and subdirectories.