Learn from downstream projects to customize processes to run Elm program
Transfer the learnings from the recent evolution of the BotLab client: Separate the compilation of the JavaScript code from the startup of the JS engine.
This week, the BotLab project introduced a new implementation for reducing agent events to an approximation of lazy evaluation. It uses an approach based on iterative probing. In that application, we perform probing on a separate JS engine to reduce the application's response times. (As the JS engine only allows a single thread at any given time)
With that new design, writing the composition log is not synced with the processing in the primary JS engine but happens after the reduction and probing in a secondary JS engine.