Speed up Elm programs by selectively disabling partial application
Discover a new way to reduce program code and runtime expenses: Change the types of functions for those we never use with partial application anyway: Reduce the parameter count to one by wrapping in a tuple/triple. This refactoring makes it easier for the compiler to reduce program code and optimize: It prevents our compiler from applying the generic wrapping that is flexible enough to support a partial application. This commit only changes one (non-exposed) helper function in the core libraries where this optimization is possible.
In the interactive scenarios using `List.range`, this change of `rangeHelp` significantly reduced the amount of work at runtime.
We could also change the compiler to detect where such a transformation is applicable and apply it automatically.