Expand state shim to inform if function application changed app state
Some projects use this information to decide whether to persist the originating event. One example is rate-limiting in a web server application: When a client floods the server with HTTP requests when the server rejects 99% of them because of a rate-limiting scheme, it may be better not to persist these events in the event log.