Fix reporting failed assembly reference resolutions for Chinese locales
Fix a bug that caused the error messages missing the details of failed assembly hash reference resolutions for volatile processes on environments with Chinese locale: Make the heuristic to find a corresponding resolution report for a given error message more robust.
Also, refactor the implementation of assembly reference resolution for volatile process programs for testability.