Fix non-strict evaluation and termination of recursive functions
Fix non-strict evaluation for the branches in a conditional expression compiled to CSharp. The termination of recursive functions depends on this. Introduce tracking of types of emitted expressions. For now, distinguish only between play values and the type wrapping into `Result<,>` Use this newly tracked information about types to change the mapping of intermediate results (`AndThen`) to happen as late and lazy as possible.
Since mapping the `Result` instances now happens closer to the leaves, remove the rootward propagation of local dependencies and the emission of statements for these accordingly.