Remove dependency on default .NET extraction process on program startup
Complete refactoring the integration of native dependencies to move the remaining ones that triggered the default .NET extraction process on program startup.

In earlier versions, on MacOS environments, we observed crashes with error messages like the following:

```
Failure processing application bundle. Failed to determine location for extracting embedded files. DOTNET_BUNDLE_EXTRACT_BASE_DIR is not set, and a read-write cache directory couldn't be created.
```

The changes in this commit complete the refactoring around native dependencies to avoid these crashes.

+ Adapt consumers of LibGit2Sharp to invoke the setup of native dependencies if necessary.
+ Refactor moving common functionality around the setup of native dependencies in a shared module.
+ Adapt the build automation to disable the extraction and adapt the collection of files for releases to avoid noise in the downloads for users.
+ Expand the testing automation to adapt to the recent observations of problems with integrating native dependencies: Add the `self-test` command on the command-line interface to run tests for functionality that depends on native dependencies, such as libgit2 and ClearScript.V8. Expand the automated checks to integrate these new tests with the same builds published with releases (single-file!).