Support deterministic profiling making the caching configurable
Enable deterministic counting of evaluation steps in program execution by moving the non-deterministic cache out of the core into a separate module. The number of evaluation steps serves as a proxy to compare the runtime efficiency of different low-level representations of the same Elm program and thus helps compare compilers.
This commit introduces a way to disable the current (non-deterministic) cache via the command-line interface.