Fix display of syntax errors and clean up compilation
Fix the display of syntax errors and clean up the implementation of lowering and compilation in the process.
The previous version did not invoke 'elm make' when lowering failed. That means we did not get the error messages from 'elm make' in that case.
This commit changes the compilation to invoke 'elm make' independent of whether lowering succeeded, so we get the full error messages also when lowering fails.