Simplify compilation introducing Elm Function packaging convention
Simplify program codes for partial application of Elm functions with a convention to package them with a description including the number of expected parameters.
Adapt the consuming site (where the next argument is applied) to check if the value to be interpreted as a function follows that pattern. If so, apply the next argument using a generic intermediate 'DecodeAndEval' expression.
Switching to this generic packaging of Elm functions allows us to eliminate the more specific templates in the compiler program code and overcome the earlier limits of function parameter count.