Fix consistency between functions to encode Pine expression as value
To fix equality checks for expressions that rely on the hashes of the value encodings.
In the previous version, the equality checks used with the `decodeExpressionOverrides` dictionary could fail because of these differences in how we encoded expressions as values.
Some encoding functions use an Elm `Dict`, resulting in the alphabetically sorted reordering of the fields, independent of the original order in the source code. Other encoding functions used a custom, different order.
This commit changes the order of the fields in the various encoding functions to match each other.
For the long-term encoding of expressions, we might use a format easily representable in a file system tree via the standard file tree encoding. Therefore it seems better to order the fields as in the common file tree encoding.