Add native v. process to avoid limitations of existing volatile process
Expand the volatile process model to avoid limitations observed with the existing CSharp-based flavor. That variant could fail script compilation in variants of Elm-Time built with the PublishSingleSingle option. For a discussion of this issue, see https://github.com/dotnet/roslyn/issues/50719

The new native flavor of volatile processes supports pre-built binaries as program code and uses the standard input/output/error streams as the interface. This commit adds a new version of the `WebService` platform to the compiler to support the new volatile process and an example Elm app and automated test using the new API.

We will probably use a WASM/WASI variant in most cases in the future since that promises a simple solution for containment. For the apps that prefer native binaries, we might add a container model and use something like docker to solve containment. That would also allow placing multiple volatile processes in the same container, which may help integrate legacy components.