Reduce compilation times and emitted program size by reducing inlining
The extent of inlining we used earlier led to the quick growth of emitted code, approaching exponential growth, as inlining a function can multiply the representations of an argument. For example, testing with the core Dict module from (https://github.com/elm/core/blob/84f38891468e8e153fc85a9b63bdafd81b24664e/src/Dict.elm) revealed that the 'insertHelp' declaration in that module grew to more than ten megabytes in size this way.

The reduction of inlining implemented with this commit reduced the emitted module size to less than a third of the previous size.

While working on the change leading up to this commit, I discovered there seems to be a bug that is currently masked by the inlining transformation in `emitExpressionInDeclarationBlock`

The balance of inlining emitted here is not meant to be optimal for runtime, as it's just an intermediate representation to be further processed. In most cases, the Pine runtime will apply additional selective inlining to optimize for runtime.