Speed up compilation of Elm core modules - selective expose
Speed up the compilation of the Elm core modules by implementing selective exposure according to the module text: Only emit code for the subset of function declarations we need to expose.
Another benefit of this change is that the emitted modules are smaller now, saving time in all following processing and compilation stages.