Clean up and organize scenarios for automated testing🧹
Remove some obsolete scenarios and move scenarios from the Elm module to the general reference collection of Elm Interactive test scenarios.
The reference scenarios are maintained in the 'test-and-train' directory. These are the tests that run automatically for all commits and pull requests.
The tests in the `ElmInteractiveTests.elm` module provide additional entry points for inspection using `elm-test-rs`