Order Elm modules by import relations for incremental compilation
Adapt the .NET side of Elm app compilation to ensure we give the Elm module texts to the Elm compiler ordered with imported modules before importing modules.
Some time ago, we switched the integration of the compiler in the Elm interactive to invoke the part currently residing in a JS engine incrementally for individual Elm modules instead of handing over the whole list of modules at once.
The side of the compiler implemented in Elm already sorts modules by import relations, but because of the batching mentioned above, it does not always get to see the whole set of modules at once, to begin with.
However, that part of the compiler also depends on the correct ordering, failing compilation if the current batch contains a reference into a module yet compiled. So compilation would fail for some context apps, depending on the order of modules there (possibly influenced by alphabetical sorting by their file names)
This commit adds sorting of the Elm modules before batching to avoid compilation failing because of incorrect compilation order.